Create load user by username and rate manga methods in User Service

+++ src/main/java/com/NowakArtur97/WorldOfManga/model/User.java
@@ -68,7 +68,7 @@ public class User {
 		this.getRoles().add(role);
 	}
 
-	public void addMangaRating(Manga manga, Double rating) {
+	public void addMangaRating(Manga manga, Long rating) {
 
 		MangaRating mangaRating = new MangaRating(manga, this, rating);
